WWMI: Option to not configure Engine.ini
Added a checkbox in the General Settings section of Wuthering Waves
to enable or disable automatic configuration of the Engine.ini file.
This option is positioned above the Engine Settings section.

Additionally, the "Open Engine.ini" button has been moved from the
Tweaks section to sit beside the new checkbox for better accessibility
and context.